// src/app/interfaces/recordatorio.ts
export interface Recordatorio {
  id?: number;
  asignacionId?: number;
  ejecutivoId?: number;
  socioId?: number;
  creditoId?: number;
  fecha?: string;
  hora?: string;
  tipoActividad?: string;
  numeroTelefono?: string;
  correo?: string;
  tipoMetodo?: string;
  direccion?: string;
  comentario?: string;
  observacion?: string;
  condicion?: string;
  fechaCreacion?: string;
}